Refactor OpentelemetryOban test setup
