--- instrumentation/opentelemetry_oban/docker-compose.yml
@@ -1,11 +1,0 @@
-version: "3.7"
-
-services:
-  postgres:
-    image: postgres:17.2
-    environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
-      - POSTGRES_DB=opentelemetry_oban_test
-    ports:
-      - 5432:5432
