+++ instrumentation/opentelemetry_oban/test/support/jobs/test_job.ex
@@ -1,0 +1,8 @@
+defmodule TestJob do
+  use Oban.Worker, queue: :events, max_attempts: 1
+
+  @impl Oban.Worker
+  def perform(_job) do
+    :ok
+  end
+end
