+++ instrumentation/opentelemetry_oban/test/support/jobs/test_job_that_returns_error.ex
@@ -1,0 +1,8 @@
+defmodule TestJobThatReturnsError do
+  use Oban.Worker, queue: :events, max_attempts: 1
+
+  @impl Oban.Worker
+  def perform(_job) do
+    {:error, :something}
+  end
+end
