+++ instrumentation/opentelemetry_oban/test/support/jobs/test_job_that_throws_exception.ex
@@ -1,0 +1,10 @@
+defmodule TestJobThatThrowsException do
+  use Oban.Worker, queue: :events, max_attempts: 1
+
+  @impl Oban.Worker
+  def perform(_job) do
+    raise %UndefinedFunctionError{
+      message: "function Some.error/0 is undefined (module Some is not available)"
+    }
+  end
+end
