+++ instrumentation/opentelemetry_oban/test/support/jobs/test_job_with_inner_span.ex
@@ -1,0 +1,11 @@
+defmodule TestJobWithInnerSpan do
+  use Oban.Worker, queue: :events, max_attempts: 1
+  require OpenTelemetry.Tracer
+
+  @impl Oban.Worker
+  def perform(_job) do
+    OpenTelemetry.Tracer.with_span "span inside the job" do
+      :ok
+    end
+  end
+end
